hafizhelper-backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
black==24.1.1
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
coverage==7.4.1
cryptography==42.0.5
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-safedelete==1.3.3
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==22.7.0
filelock==3.13.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
httplib2==0.22.0
identify==2.5.33
isort==5.13.2
msgpack==1.0.8
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pip-autoremove==0.10.0
platformdirs==4.2.0
pre-commit==3.6.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
whitenoise==6.6.0
